Feature / Add parameter to disable deletion before import

+++ lib/csv_fast_importer.rb
@@ -12,10 +12,12 @@ class CsvFastImporter
 
     row_index = 0
     within_transaction_if(configuration.transactional?) do
-      if configuration.truncate?
-        sql_connection.execute "TRUNCATE TABLE \"#{sql_table}\""
-      else
-        sql_connection.execute "DELETE FROM \"#{sql_table}\""
+      if configuration.deletion?
+        if configuration.truncate?
+          sql_connection.execute "TRUNCATE TABLE \"#{sql_table}\""
+        else
+          sql_connection.execute "DELETE FROM \"#{sql_table}\""
+        end
       end
       sql_connection.raw_connection.copy_data <<-SQL do
         COPY "#{sql_table}" (#{sql_columns})
